Rename action files to use camel case instead of kebab case
